Move aggregted_body to its own file move all block code to the own folder
